(#21) :bug: Remove --allow-multiple for uninstall (#32)
* :recycle: Rework logic to use allow-multiple as needed

In some instances, allow-multiple is *required*,
specifically if there are multiple package versions installed.

Fix for this to work correctly seems to require:
- Apply allow_multiple if and only if multiple versions are installed
- DO NOT apply allow_multiple when uninstalling the only
  version of a package that is installed.

* :white_check_mark: Update test

The test for uninstall specific versions when multiple are installed
seems to be targeting two different packages potentially.

Both checks should be looking for the same package name.

* :construction: ensure pip is installed in the server

Apparently pip is missing, somehow. Not really sure how/why on that one.