:bug: (#35) Fix state:absent for specific versions
Previously, specifying a direct version would have choco error out with
either state: absent or state: reinstalled, which we don't want.

Fixed by allowing the version value provided to be passed through to
the choco list command, which will allow us to filter down to only
target installed _versions_ of packages, if desired.
When state:absent is supplied along with a target version, module will
no-op when that version is not already installed.