(maint) Add files for locally testing Chocolatey
This adds a Vagrantfile w/ instructions to allow local testing of new
versions of Chocolatey via the Ansible collection.

The purpose of this is to allow us to double-check new Chocolatey beta
versions to ensure that there is no critical functionality missing from
the Ansible collection, as well as ensuring that new Chocolatey versions
don't break any existing critical functionality.

[skip ci]

This addition doesn't affect the collection itself, so we can safely
skip CI.