(#47) :white_check_mark: Fix test cleanup
The win_chocolatey_source tests re-added the default source with a
capital C in the name, which doesn't match the normal default (all
lowercase).
This causes other tests to fail on a rerun, as the source name is now
different than expected.

Fix is to ensure we re-add the source name in lowercase.