(#40) :bug: Don't warn on explicit Choco install (#43)
Previously, if Chocolatey was not already on the system, the collection
would emit a warning when installing Chocolatey, even if the current
task was specifically to install Chocolatey anyway.

This change adds a parameter to Install-Chocolatey which allows the
warning to be skipped. This parameter is supplied only if the task
explicitly requests Chocolatey itself to be installed.