:rocket: (#57) Build collection before testing
Previously the collection would be built during the test phase, but this
caused issues with the matrix strategy, so consolidated that into a
separate stage and have that done ahead of time so that we only need to
publish the collection artifact once.

Also reworked some of the logic for how the inventory file is added in
the invoke script in CI to account for the change in build process.