(#58) Save chocolatey install script to temp file
Running scripts by piping over stdin directly to powershell can be
pretty problematic for a few reasons, one of which is that certain
control flow statements don't quite work correctly.

Instead of piping directly to stdin, we can save the script to a
temporary file and have powershell invoke that directly instead.

Additionally includes a fix for the underlying issue in #58 which
occurred due to usage of Expand-Archive in the install script which was
trying to draw a progress bar in our non-interactive context.