(#15) Allow additional args for choco
This allows users to specify arguments that are not explicity supported
by the collection, for example any options that require the chocolatey
licensed extension to be installed.

This also gives us a workaround / stop-gap for any additional arguments
in open source Chocolatey in future, allowing users to specify them
manually until the collection is updated to include those options.