(maint) Fix sanity test issues
A few sanity tests started failing, not sure when, just missing metadata
in the new argument spec. Think it makes enough sense to fix them
here rather than wait.

As part of this, we need to be absolutely sure pwsh is installed on the
pipeline runner, so that Ansible can run the appropriate sanity tests.