(#17) Add TLS option for bootstrapping
Previously, the TLS configuration was set to allow both TLS 1.1 and
TLS 1.2.

With this change, users can explicitly set the TLS versions they want to
allow during bootstrapping. The default settings are to allow TLS 1.2
and 1.3, according to the TLS versions available on the client.

Also added some tests to verify the new behaviour.