(#122) Consolidate bootstrap tests
The issues in testing v1 looked to be related to some bootstrap tests
left in the main test run, which are duplicates of those that run as
part of the initial test sequence before the bulk of the tests.

Removed the duplicate tests, moved some of the other potentially
state-destructive tests into the bootstrap tests, where we handle
things better.