Fix crashing when PATH contains multiple choco installations
When for some reason the path environment variable contains two installations of chocolatey this function will return a list, which causes a syntax error down the line. Since running choco.exe on the machine will also choose the first installation in the path, I think limiting the found installations to the first is the desired behavior here.