(#173) Add support for --ignore-pinned switch
The --ignore-pinned switch allows to the upgrade of pinned packages without
needing to unpin, upgrade, and then re-pin the package. It was added to
Chocolatey CLI in v2.3.0 and the win_chocolatey module will fail the task
if an unsupported version of Chocolatey CLI is detected.
Test have been added to validate this change is effective.